PHOENIX-114 Migrate Demand Frame work to Phoenix

+++ egov/egov-demand/src/main/java/org/egov/demand/interfaces/BillServiceInterface.java
@@ -106,7 +106,7 @@ public abstract class BillServiceInterface {
 	public final EgBill generateBill(Billable billObj) {
 		EgBill bill = new EgBill();
 		bill.setBillNo(billObj.getReferenceNumber());
-		bill.setBoundaryNum((billObj.getBoundaryNum()));
+		bill.setBoundaryNum(billObj.getBoundaryNum().intValue());
 		bill.setBoundaryType(billObj.getBoundaryType());
 		bill.setCitizenAddress(billObj.getBillAddress());
 		bill.setCitizenName(billObj.getBillPayee());
